Moved entities in trait to upper level
