+++ backend/libs/exchanges/history/src/entities/mod.rs
@@ -1,0 +1,3 @@
+mod status;
+
+pub use self::status::KlineFetchStatus;
