+++ backend/libs/exchanges/history/src/entities/status.rs
