+++ backend/libs/exchanges/history/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod binance;
+pub mod entities;
 pub mod traits;
