+++ backend/libs/exchanges/history/src/traits/mod.rs
@@ -1,7 +1,6 @@
-pub mod entities;
 mod fetcher;
+pub mod kline;
 mod pubsub;
-pub mod traits;
 mod writer;
 
 pub use self::fetcher::HistoryFetcher;
